+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
 
+import ru.yandex.practicum.filmorate.custom_exceptions.IncorrectParameterException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
@@ -44,6 +45,8 @@ public class UserController {
     @PutMapping("/{id}/friends/{friendId}")
     public void addFriend(@PathVariable int id,
                           @PathVariable int friendId) {
+        if (id <= 0) throw new IncorrectParameterException("id");
+        if (friendId <= 0) throw new IncorrectParameterException("friendId");
         userService.addFriend(id, friendId);
     }
 
@@ -55,6 +58,8 @@ public class UserController {
     @DeleteMapping("/{id}/friends/{friendId}")
     public void deleteFriend(@PathVariable int id,
                              @PathVariable int friendId) {
+        if (id <= 0) throw new IncorrectParameterException("id");
+        if (friendId <= 0) throw new IncorrectParameterException("friendId");
         userService.deleteFriend(id, friendId);
     }
 
@@ -65,17 +70,21 @@ public class UserController {
 
     @GetMapping("/{id}")
     public User findGyId(@PathVariable int id) {
+        if (id <= 0) throw new IncorrectParameterException("id");
         return userService.getById(id);
     }
 
     @GetMapping("/{id}/friends")
     public List<User> findFriends(@PathVariable int id) {
+        if (id <= 0) throw new IncorrectParameterException("id");
         return userService.getFriends(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
     public List<User> getCommonFriends(@PathVariable int id,
                                       @PathVariable int otherId) {
+        if (id <= 0) throw new IncorrectParameterException("id");
+        if (otherId <= 0) throw new IncorrectParameterException("otherId");
         return userService.getCommonFriends(id, otherId);
     }
 }
